Remove superfluous module hierarchy, delete superfluous test
